:recycle: Move BTC send form tests into its file
